+++ styles/globals.css
@@ -188,7 +188,7 @@ h1 {
 
 @media (min-width: 1280px) {
   header {
-    padding: 0 200px;
+    padding: 0 100px;
   }
 
   .wrapper {
